Fixed accuracy "exactly" with diacritics (#29)
When searching for "ipsu" with "exactly" accuracy, you would expect only this exact standalone word to be marked. However, when searching with diacritics also "ipsü" inside "ipsüm" will be marked. This happens due to the fact that the word boundary will only match ASCII characters (and ü is a unicode character). This is now solved.