Refactored noHeading in articles and navigation
Previously "noHeading" was specified in the TOC itself. This prevents that some articles should be included without a heading (because they use the heading of the parent), but sub-articles still should have one. This is now fixed due to the fact that "heading: false" is specified on each article rather than in the TOC.